requirements.txt
altgraph==0.17.4
mido==1.3.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-rtmidi==1.5.8
pywin32-ctypes==0.2.3
